feat: enhance availability calendar
